Refactor and create common abstraction for Storage Providrs
